requirements.txt
langgraph
langchain-core
langchain_google_genai
langchain_openai
playwright
html2text
pandas
colorama
python-dotenv
bs4
